test(signals): fold reward-risk-engine-branch-coverage.test.ts into its siblings (#8578)
reward-risk-engine-branch-coverage.test.ts was a #2281 Codecov bolt-on
testing buildRepoRewardRisk / rewardRiskFreshnessInternals /
buildContributorRewardRiskStrategy from its own disconnected file. Move
each case to the sibling that already owns the function under test:

- bestFitLabels tie-break -> reward-risk-freshness.test.ts's existing
  "bestFitLabels keyword anchoring" describe (reused its pick() idiom)
- reviewChurnRisk (high/medium) + maintainer-cut readiness ->
  reward-risk-freshness.test.ts, new describe block (buildRepoRewardRisk)
- buildContributorRewardRiskStrategy tie-break -> reward-risk-reports.test.ts,
  new describe block (the only reward-risk sibling besides the signals-v2/
  signals-coverage family, which stays out of scope per #8574/#8576)

Verified zero coverage regression: packages/loopover-engine/src/reward-risk.ts
(259/303 branches, 152/164 lines) and packages/loopover-engine/src/signals/engine.ts
(441/1973 branches, 437/1351 lines) are identical before and after.
